scripts/chatbot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pydantic==2.9.0
psycopg2-binary==2.9.9
